+++ src/lib.rs
@@ -31,7 +31,7 @@ pub use crate::client::{
     Session, SessionReceiver, SessionEvent, SessionResp, ExitSignal,
     PtyRequest, PtyTerminalModes, WindowChange,
 };
-pub use crate::client::{Tunnel, TunnelReceiver, TunnelEvent};
+pub use crate::client::{Tunnel, TunnelReceiver, TunnelEvent, TunnelReader, TunnelWriter, TunnelStream};
 pub use crate::codec::{PacketEncode, PacketDecode};
 pub use crate::error::{Result, Error, AlgoNegotiateError, DisconnectError, ChannelOpenError};